style: code styling
