+++ app/views/__init__.py
@@ -2,4 +2,6 @@ from .user_response                 import *
 from .user_token_response           import *
 from .user_did_response             import *
 from .cover_letter_response         import *
+from .cover_letter_id_response      import *
 from .position_response             import *
+from .position_id_response          import *
