+++ app/views/cover_letter_response.py
@@ -1,4 +1,3 @@
-from app.serializers.CL_serializer import CLDetailSerializer
 from .. import errors
 from ..utils import cryptoutils, modelutils
 from ..models import CLData
@@ -10,25 +9,6 @@ from rest_framework.response import Response
 
 class CoverLetterResponse(APIView):
 
-    @staticmethod
-    def gen_post_response(code=status.HTTP_201_CREATED, err=None):
-        return Response(
-            {
-                "error": err,
-            },
-            status=code,
-        )
-
-    @staticmethod
-    def gen_get_response(cl_list, code=status.HTTP_200_OK, err=None):
-        return Response(
-            {
-                "error": err,
-                "cover_letters": cl_list,
-            },
-            status=code,
-        )
-
     """
     [POST] /api/cover-letter
     @PathVariable: nil
@@ -38,6 +18,15 @@ class CoverLetterResponse(APIView):
         cover-letter:   string
     }
     """
+    @staticmethod
+    def gen_post_response(code=status.HTTP_201_CREATED, err=None):
+        return Response(
+            {
+                "error": err,
+            },
+            status=code,
+        )
+
     def post(self, request):
         try:
             # Check login
@@ -81,6 +70,16 @@ class CoverLetterResponse(APIView):
     @RequestParam: nil
     @RequestBody: nil
     """
+    @staticmethod
+    def gen_get_response(cl_list, code=status.HTTP_200_OK, err=None):
+        return Response(
+            {
+                "error": err,
+                "cover_letters": cl_list,
+            },
+            status=code,
+        )
+
     def get(self, request):
         try:
             # Check login
@@ -96,49 +95,6 @@ class CoverLetterResponse(APIView):
             serializer = CLMinimumSerializer(
                 CLData.objects.filter(owner = did), # type: ignore
                 many=True
-            )
-            return self.gen_get_response(serializer.data)
-
-        # Handle all known error
-        except errors.BaseError as err:
-            return err.gen_response()
-
-        # Unknown error
-        except Exception as err:
-            return errors.UnhandledError(err).gen_response()
-
-class CoverLetterIDResponse(APIView):
-
-    @staticmethod
-    def gen_get_response(cl, code=status.HTTP_200_OK, err=None):
-        return Response(
-            {
-                "error": err,
-                "cover_letter": cl,
-            },
-            status=code,
-        )
-
-    """
-    [GET] /api/cover-letter/:cl_id
-    @PathVariable: nil
-    @RequestParam: nil
-    @RequestBody: nil
-    """
-    def get(self, request, cl_id):
-        try:
-            # Check login
-            if "access_token" not in request.COOKIES:
-                raise errors.AuthorizationFailedError("Access token not exists")
-            did = cryptoutils.verify_JWT(request.COOKIES["access_token"])
-
-            # Check employee
-            if not modelutils.is_employee(did):
-                raise errors.PermissionDeniedError()
-
-            # Generate serializer
-            serializer = CLDetailSerializer(
-                    CLData.objects.get(owner = did, pk=cl_id), # type: ignore
             )
             return self.gen_get_response(serializer.data)
